Added: loading bar, key bind fix, clean up of UI
I added the loading bar and it has its own script. Right now the string
'path' that it uses is hardcoded in as it is a public string. Key
bindings are being read and loaded at on enabled in the player manager
and game manager scripts. Cleaned up the UI a little bit. Most of the
menus look a lot better except for the key bindings menu. I want to go
over it with you guys to see how to design it.

BUG: I have encountered a problem with the spawning of a player at the
beggining of the round. I was not having this problem consitently so I
am having a hard time trouble shooting it. I would like to go over the
scripts that start the round. There are quite a few different scripts
talking to each other and many are networked so that may be the issue.

Another small bug is, I have not yet implemented the part that changes
the key bindings button text when you change the key binding. I will be
working on that next tonight and should have it ready to go by tomorrow
when I want to meet with you guys.