Bug Fixes
Fixed:

-You can not click on any of the drop downs in the display settings.

-You can't check the full screen option in the display settings.

-You can't change the motion blur from false to true.

-You can choose true or false for ambient occlusion but it does not
save.

-Selecting full res, sets the resolution to eighth res and vice versa

-The Audio button doesn't bring up anything.

-The slider in the controls settings should say that it is sensitivity.

I had added the ui background and that prevented anyone from clicking on
the dropdowns. That was a last minute untested add so I apologize. I
updated the save files.

I am going to be working on the key binding bugs next. I already have
ideas for most of them and will start implementation now. I do still
have to bring over a new scene for the loading screen.

I will find better UI art and use that for the menu, temporarily. I also
plan on adding a rotating model of our character or a model of the
character doing some idle animations.